Automate the frame processing pipeline in the scheduler
