+++ src/transmission/forms/forms.py
@@ -42,4 +42,3 @@ class RemoveJob(forms.Form):
 
     sat = forms.ChoiceField(choices=satellites, widget=forms.Select, label='Satellite')
     job_type = forms.ChoiceField(choices=jobs, widget=forms.Select, label='Job Type')
-
